+++ actstream/models.py
@@ -4,30 +4,15 @@ from django.conf import settings
 from django.contrib.contenttypes.fields import GenericForeignKey
 from django.db import models
 from django.utils.translation import ugettext as _
-from django.utils.encoding import python_2_unicode_compatible
 from django.utils.timesince import timesince as djtimesince
 from django.contrib.contenttypes.models import ContentType
-
-try:
-    from django.urls import reverse
-except ImportError:
-    from django.core.urlresolvers import reverse
-
-try:
-    from django.utils import timezone
-    now = timezone.now
-except ImportError:
-    from datetime import datetime
-    now = datetime.now
-
+from django.urls import reverse
+from django.utils.timezone import now
 
 from actstream import settings as actstream_settings
 from actstream.managers import FollowManager
 
-now = timezone.now
 
-
-@python_2_unicode_compatible
 class Follow(models.Model):
     """
     Lets a user follow the activities of any specific actor
@@ -54,10 +39,9 @@ class Follow(models.Model):
         unique_together = ('user', 'content_type', 'object_id', 'flag')
 
     def __str__(self):
-        return '%s -> %s : %s' % (self.user, self.follow_object, self.flag)
+        return '{} -> {} : {}'.format(self.user, self.follow_object, self.flag)
 
 
-@python_2_unicode_compatible
 class Action(models.Model):
     """
     Action model describing the actor acting out a verb (on an optional