+++ docs/source/changelog.rst
@@ -3,17 +3,27 @@
 Changelog
 =========
 
+0.8.0
+-----
+  - Drop django 2.0
+  - Drop python 3.4 and 2.7
+  - Add django 2.2
+  - mysql jsonfied
+  - numerious bug fixes
+
+
 0.7.0
 -----
-
+  - Remove django less then 1.11 compatibiity
   - Remove django less then 1.8 compatibility
   - Adjusted URLs with new versions of django
   - Improved tox and travis config
   - use render function instead of depricated render_to_response
   - upgrade package depedencies
-  - django 2.0 support
-  - SubQuery in follow query
-  
+  - django 1.11, 2.0 and 2.1 support added
+  - Code clean ups
+  - Add optional support for django-mysql's JSONField
+
 0.6.5
 -----